% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_classification.R
\name{AUC}
\alias{AUC}
\title{AUC of ROC graph}
\usage{
AUC(fitval, y)
}
\arguments{
\item{fitval}{An numeric array of diagnostic score.}

\item{y}{An array of equal length of score, containing the class of the observations.}
}
\value{
The AUC of an ROC graph
}
\description{
AUC of ROC graph
}
