% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{count_cases}
\alias{count_cases}
\title{Count the number of crimes in a given time frame and location level}
\usage{
count_cases(
  df,
  date_start = NULL,
  date_end = NULL,
  location_level = NULL,
  date_level = "month"
)
}
\arguments{
\item{df}{Data frame}

\item{date_start}{Date to begin the count}

\item{date_end}{Date to end the count}

\item{location_level}{Column name of the relevant location level from the data set.}

\item{date_level}{String specifying the date level required. Options are "day", "week" or "month".}
}
\value{
Data frame containing the relevant count data
}
\description{
Count the number of crimes in a given time frame and location level
}
