% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{get_indicators}
\alias{get_indicators}
\title{Load socioecononic indicators data from the Chicago database}
\usage{
get_indicators()
}
\value{
The socioeconomic data frame from the Chicago databases
}
\description{
Load socioecononic indicators data from the Chicago database
}
