% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{get_map}
\alias{get_map}
\title{Load area mapping shapefile from the Chicago database}
\usage{
get_map(map = "community_area")
}
\arguments{
\item{map}{String specifying which map to load. Options are "community_area" and "district".}
}
\value{
The requested map as a shapefile consisting of geometries and the relevant area coding
}
\description{
Load area mapping shapefile from the Chicago database
}
