% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_classification.R
\name{ggplotROC}
\alias{ggplotROC}
\title{ROC plot function}
\usage{
ggplotROC(fitval, y)
}
\arguments{
\item{fitval}{An numeric array of diagnostic score.}

\item{y}{An array of equal length of score, containing the class of the observations.}
}
\value{
A ggplot2 plot
}
\description{
ROC plot function
}
