% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_crimes_API}
\alias{load_crimes_API}
\title{Load chicago crime data using Socrata API.}
\usage{
load_crimes_API()
}
\arguments{
\item{filepath}{path to the processed csv file.}
}
\value{
tibble data frame of Chicago Crime data.
}
\description{
Returns a data frame of Chicago Crime data from specified year or range of years,
or returns entire dataset from 2001 to present if no years are specified.
}
