% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_crimes_csv}
\alias{load_crimes_csv}
\title{Load chicago crime data from processed csv.}
\usage{
load_crimes_csv(filepath)
}
\arguments{
\item{filepath}{path to the processed csv file.}
}
\value{
tibble data frame of Chicago Crime data.
}
\description{
Load chicago crime data from processed csv.
}
