% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{long_to_short}
\alias{long_to_short}
\title{Converts long name format with spaces to short name format with underscores}
\usage{
long_to_short(string)
}
\arguments{
\item{string}{String to be converted}
}
\value{
the string, with underscores and uncapitalised
}
\description{
Converts long name format with spaces to short name format with underscores
}
