% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{long_variables}
\alias{long_variables}
\title{Renames data variable names to long format, with spaces, from short, with underscores}
\usage{
long_variables(data)
}
\arguments{
\item{data}{The data set with the variables to be renamed}
}
\value{
data frame with long variable names
}
\description{
Renames data variable names to long format, with spaces, from short, with underscores
}
