% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass_classification.R
\name{mnlr_cv_indexed}
\alias{mnlr_cv_indexed}
\title{Indexed cross-validation for multinomial regression}
\usage{
mnlr_cv_indexed(X, y, index)
}
\arguments{
\item{X}{Data to be used, not including the response variable or any variables not for use in model.}

\item{y}{Response variable, as a factor.}

\item{index}{The indices of the rows to be used as test data.}
}
\value{
A confusion matrix, see `?confusionMatrix`
}
\description{
Performs cross-validation computing metrics using specified indexes of the data set.
}
