% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass_classification.R
\name{mnlr_kfold_cv.list}
\alias{mnlr_kfold_cv.list}
\title{K-fold cross-validation metrics for multinomial regression, in list format.}
\usage{
mnlr_kfold_cv.list(X, y, k, n_reps = 5, metrics)
}
\arguments{
\item{X}{Data to be used, not including the response variable or any variables not for use in model.}

\item{y}{Response variable, as a factor.}

\item{k}{Number of folds.}

\item{n_reps}{Number of repeats.}

\item{metrics}{Vector of strings with metrics to be obtained. Overall accuracy, no information rate (NIR) and the p value for accuracy > NIR will always be returned. Options are "Sensitivity", "Specificity", "Pos Pred Value", "Neg Pred Value", "Precision", "Recall", "F1", "Prevalence", "Detection Rate","Detection Prevalence" and "Balanced Accuracy". At least one must be given and must be in vector format.}
}
\value{
List containing two elements: a list of length equal to that of `metrics` + 3, with each element containing a list or vector of the mean of the metric at each repeat, averaged over `k` folds, and a list of length equal to `n_reps`, containing the metrics for each of the `k` folds.
}
\description{
K-fold cross-validation metrics for multinomial regression, in list format.
}
