% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass_classification.R
\name{mnlr_kfold_cv}
\alias{mnlr_kfold_cv}
\title{K-fold cross-validation for multinomial regression}
\usage{
mnlr_kfold_cv(X, y, k, n_reps = 5, metrics)
}
\arguments{
\item{X}{Data to be used, not including the response variable or any variables not for use in model.}

\item{y}{Response variable, as a factor.}

\item{k}{Number of folds.}

\item{n_reps}{Number of repeats.}

\item{metrics}{Vector of strings with metrics to be obtained. Overall accuracy will always be returned. Options are "Sensitivity", "Specificity", "Pos Pred Value"       "Neg Pred Value", "Precision", "Recall", "F1", "Prevalence", "Detection Rate","Detection Prevalence" and "Balanced Accuracy".}
}
\value{
List of length equal to that of `metrics` + 1, with each element containing a list or vector of the mean of the metric at each repeat, averaged over `k` folds.
}
\description{
K-fold cross-validation for multinomial regression
}
