% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{othering}
\alias{othering}
\title{Convert less common levels to "OTHER", to be used when `Primary Type` or `Location Description` is present to reduce category numbers.}
\usage{
othering(factor_vec, threshold, print_summary = FALSE)
}
\arguments{
\item{factor_vec}{Factor vector to be otherised.}

\item{threshold}{Threshold count below which will be converted to other.}

\item{print_summary}{Boolean indicating if a summary of the operations should be printed. Default is FALSE.}
}
\value{
Otherised factor vector
}
\description{
Convert less common levels to "OTHER", to be used when `Primary Type` or `Location Description` is present to reduce category numbers.
}
