% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_factors}
\alias{plot_factors}
\title{Code to produce bar plots of factor variables}
\usage{
plot_factors(data, column, threshold = NULL)
}
\arguments{
\item{data}{Data to be plotted}

\item{column}{Variable to be plotted}

\item{threshold}{Threshold value to be shown on plot. If NULL, no value shown. Default is NULL.}
}
\value{
Bar plot
}
\description{
Code to produce bar plots of factor variables
}
