% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_analysis.R
\name{plot_heat_map}
\alias{plot_heat_map}
\title{Producing a heatmap based on a variable}
\usage{
plot_heat_map(data, variable, legend.title = variable, trans = "identity")
}
\arguments{
\item{data}{A `data.frame` containing where the shapefiles are contained in the `geometry` column}

\item{variable}{A string of the column name of the variable to be used to create the heat map}

\item{legend.title}{A string of the title of the legend to be included in the plot}

\item{trans}{A string indicating the transformation to be applied to the scale (options are "asn", "atanh", "boxcox", "date", "exp", "hms", "identity", "log", "log10", "log1p", "log2", "logit", "modulus", "probability", "probit", "pseudo_log", "reciprocal", "reverse", "sqrt" and "time")}
}
\value{
A `ggplot` object with a heat map of the shapefiles in `data` with the fill colour determined by `variable`
}
\description{
Producing a heatmap based on a variable
}
