% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_analysis.R
\name{plot_heat_map_d}
\alias{plot_heat_map_d}
\title{Producing a heatmap based on a discrete variable}
\usage{
plot_heat_map_d(data, variable, legend.title = variable)
}
\arguments{
\item{data}{A `data.frame` containing where the shapefiles are contained in the `geometry` column}

\item{variable}{A string of the column name of the variable to be used to create the heat map}

\item{legend.title}{A string of the title of the legend to be included in the plot}
}
\value{
A `ggplot` object with a heat map of the shapefiles in `data` with the fill colour determined by `variable`
}
\description{
Producing a heatmap based on a discrete variable
}
