% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{report_table}
\alias{report_table}
\title{Generate table for report}
\usage{
report_table(data, align)
}
\arguments{
\item{data}{tibble containing data to be tabulated}

\item{align}{String vector with ncol(data) elements, specifying the alignments of the columns. String options are "l", "c" and "r".}
}
\value{
knitr kable with latex format
}
\description{
Generate table for report
}
