% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{short_to_long}
\alias{short_to_long}
\title{Converts short name format with underscores to long name format}
\usage{
short_to_long(string)
}
\arguments{
\item{string}{String to be converted}
}
\value{
the string, with underscores replaced by spaces and capitalised
}
\description{
Converts short name format with underscores to long name format
}
