% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{short_variables}
\alias{short_variables}
\title{Renames data variable names to short format, with underscores, from long, with spaces.}
\usage{
short_variables(data)
}
\arguments{
\item{data}{The data set with the variables to be renamed}
}
\value{
data frame with short variable names
}
\description{
Renames data variable names to short format, with underscores, from long, with spaces.
}
