% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass_classification.R
\name{significance_test}
\alias{significance_test}
\title{Significance test for features}
\usage{
significance_test(model, alpha = 0.05)
}
\arguments{
\item{model}{a multinomial logistic regression model}

\item{alpha}{numeric significance level}
}
\value{
list containing z statistics, p values and indicators of significance.
}
\description{
Tests if features are significant in a multinomial logistic regression model using 2 tail z tests
}
