% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_crimes_API}
\alias{load_crimes_API}
\title{Load chicago crime data using Socrata API.}
\usage{
load_crimes_API(year = "2019")
}
\arguments{
\item{year}{The year, between 2001 and present, of the data to be extracted. If NULL returns all data from 2001 to present. Default is "2019".}
}
\value{
tibble data frame of Chicago Crime data from the specified year, with long variable names.
}
\description{
Returns a data frame of Chicago Crime data, optionally from specified year.
}
