% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass_classification.R
\name{mean_repeat_metrics}
\alias{mean_repeat_metrics}
\title{Calculate the means across folds of performance metrics}
\usage{
mean_repeat_metrics(results, type)
}
\arguments{
\item{results}{Output of `mnlr_kfold_cv.df`}

\item{type}{String specifying which metric type to extract. Options are "class" or "overall". "class" should only be used for results containing exactly the by class metrics of Sensitivity, Specificty and Balanced Accuracy.}
}
\value{
A tibble containing the metrics, averages over all folds, for each repeat.
}
\description{
Calculate the means across folds of performance metrics
}
