% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_data}
\alias{process_data}
\title{Process Chicago Crime data}
\usage{
process_data(data)
}
\arguments{
\item{data}{Data set to process, in tibble format and with long variable names}
}
\value{
Processed data
}
\description{
Converts columns to the correct types, formatting the date variable and merging the `Primary Type` categories of "CRIM SEXUAL ASSAULT" and "CRIMINAL SEXUAL ASSAULT".
}
